+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Add `skip-get-modules` builder flag to support isolated environment executions (#6009)
   - Skip unnecessary Go binary path validation when the builder is used with `skip-compilation` and `skip-get-modules` flags (#6026)
 - Add mapstructure hook function for confmap.Unmarshaler interface (#6029)
+- Add CopyTo and MoveTo methods to primitive slices (#6044)
 
 ### 🧰 Bug fixes 🧰